+++ lua/lvim/core/autocmds.lua
@@ -103,7 +103,7 @@ function M.load_defaults()
         group = "_lvim_colorscheme",
         callback = function()
           if lvim.builtin.breadcrumbs.active then
-            require("lvim.core.builtins.breadcrumbs").get_winbar()
+            require("lvim.builtin.breadcrumbs").get_winbar()
           end
           local statusline_hl = vim.api.nvim_get_hl_by_name("StatusLine", true)
           local cursorline_hl = vim.api.nvim_get_hl_by_name("CursorLine", true)
