+++ lua/lvim/plugins.lua
@@ -24,7 +24,7 @@ local core_plugins = {
   {
     "williamboman/mason.nvim",
     config = function()
-      require("lvim.core.builtins").setup "mason"
+      require("lvim.builtin").setup "mason"
     end,
     cmd = { "Mason", "MasonInstall", "MasonUninstall", "MasonUninstallAll", "MasonLog" },
     lazy = true,
@@ -46,7 +46,7 @@ local core_plugins = {
     "nvim-telescope/telescope.nvim",
     branch = "0.1.x",
     config = function()
-      require("lvim.core.builtins").setup "telescope"
+      require("lvim.builtin").setup "telescope"
     end,
     dependencies = { "telescope-fzf-native.nvim" },
     lazy = true,
@@ -58,7 +58,7 @@ local core_plugins = {
   {
     "hrsh7th/nvim-cmp",
     config = function()
-      require("lvim.core.builtins").setup "cmp"
+      require("lvim.builtin").setup "cmp"
     end,
     event = { "InsertEnter", "CmdlineEnter" },
     dependencies = {
@@ -112,7 +112,7 @@ local core_plugins = {
     "windwp/nvim-autopairs",
     event = "InsertEnter",
     config = function()
-      require("lvim.core.builtins").setup "autopairs"
+      require("lvim.builtin").setup "autopairs"
     end,
     enabled = lvim.builtin.autopairs.active,
   },
@@ -125,7 +125,7 @@ local core_plugins = {
       local utils = require "lvim.utils"
       local path = utils.join_paths(get_runtime_dir(), "site", "pack", "lazy", "opt", "nvim-treesitter")
       vim.opt.rtp:prepend(path) -- treesitter needs to be before nvim's runtime in rtp
-      require("lvim.core.builtins").setup "treesitter"
+      require("lvim.builtin").setup "treesitter"
     end,
     cmd = {
       "TSInstall",
@@ -148,7 +148,7 @@ local core_plugins = {
   {
     "kyazdani42/nvim-tree.lua",
     config = function()
-      require("lvim.core.builtins").setup "nvimtree"
+      require("lvim.builtin").setup "nvimtree"
     end,
     enabled = lvim.builtin.nvimtree.active,
     cmd = { "NvimTreeToggle", "NvimTreeOpen", "NvimTreeFocus", "NvimTreeFindFileToggle" },
@@ -158,7 +158,7 @@ local core_plugins = {
   {
     "tamago324/lir.nvim",
     config = function()
-      require("lvim.core.builtins").setup "lir"
+      require("lvim.builtin").setup "lir"
     end,
     enabled = lvim.builtin.lir.active,
     event = "User DirOpened",
@@ -166,7 +166,7 @@ local core_plugins = {
   {
     "lewis6991/gitsigns.nvim",
     config = function()
-      require("lvim.core.builtins").setup "gitsigns"
+      require("lvim.builtin").setup "gitsigns"
     end,
     event = "User FileOpened",
     cmd = "Gitsigns",
@@ -177,7 +177,7 @@ local core_plugins = {
   {
     "folke/which-key.nvim",
     config = function()
-      require("lvim.core.builtins").setup "which-key"
+      require("lvim.builtin").setup "which-key"
     end,
     cmd = "WhichKey",
     event = "VeryLazy",
@@ -188,7 +188,7 @@ local core_plugins = {
   {
     "numToStr/Comment.nvim",
     config = function()
-      require("lvim.core.builtins").setup "comment"
+      require("lvim.builtin").setup "comment"
     end,
     keys = { { "gc", mode = { "n", "v" } }, { "gb", mode = { "n", "v" } } },
     event = "User FileOpened",
@@ -199,7 +199,7 @@ local core_plugins = {
   {
     "ahmedkhalf/project.nvim",
     config = function()
-      require("lvim.core.builtins").setup "project"
+      require("lvim.builtin").setup "project"
     end,
     enabled = lvim.builtin.project.active,
     event = "VimEnter",
@@ -219,7 +219,7 @@ local core_plugins = {
     "nvim-lualine/lualine.nvim",
     -- "Lunarvim/lualine.nvim",
     config = function()
-      require("lvim.core.builtins").setup "lualine"
+      require("lvim.builtin").setup "lualine"
     end,
     event = "VimEnter",
     enabled = lvim.builtin.lualine.active,
@@ -229,7 +229,7 @@ local core_plugins = {
   {
     "SmiteshP/nvim-navic",
     config = function()
-      require("lvim.core.builtins").setup "breadcrumbs"
+      require("lvim.builtin").setup "breadcrumbs"
     end,
     event = "User FileOpened",
     enabled = lvim.builtin.breadcrumbs.active,
@@ -238,7 +238,7 @@ local core_plugins = {
   {
     "akinsho/bufferline.nvim",
     config = function()
-      require("lvim.core.builtins").setup "bufferline"
+      require("lvim.builtin").setup "bufferline"
     end,
     branch = "main",
     event = "User FileOpened",
@@ -249,7 +249,7 @@ local core_plugins = {
   {
     "mfussenegger/nvim-dap",
     config = function()
-      require("lvim.core.builtins").setup "dap"
+      require("lvim.builtin").setup "dap"
     end,
     lazy = true,
     enabled = lvim.builtin.dap.active,
@@ -269,7 +269,7 @@ local core_plugins = {
   {
     "goolord/alpha-nvim",
     config = function()
-      require("lvim.core.builtins").setup "alpha"
+      require("lvim.builtin").setup "alpha"
     end,
     enabled = lvim.builtin.alpha.active,
     event = "VimEnter",
@@ -281,7 +281,7 @@ local core_plugins = {
     event = "VeryLazy",
     branch = "main",
     config = function()
-      require("lvim.core.builtins").setup "terminal"
+      require("lvim.builtin").setup "terminal"
     end,
     enabled = lvim.builtin.terminal.active,
   },
@@ -295,7 +295,7 @@ local core_plugins = {
   {
     "RRethy/vim-illuminate",
     config = function()
-      require("lvim.core.builtins").setup "illuminate"
+      require("lvim.builtin").setup "illuminate"
     end,
     event = "User FileOpened",
     enabled = lvim.builtin.illuminate.active,
@@ -304,7 +304,7 @@ local core_plugins = {
   {
     "lukas-reineke/indent-blankline.nvim",
     config = function()
-      require("lvim.core.builtins").setup "indentlines"
+      require("lvim.builtin").setup "indentlines"
     end,
     event = "User FileOpened",
     enabled = lvim.builtin.indentlines.active,
